# server/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CURREX API 명세",
    "description": "CURREX 서비스의 다양한 API를 테스트합니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://currex.kro.kr"
    }
  ],
  "paths": {
...
    "/api/auth/kakao/callback": {
      "get": {
        "tags": ["로그인/로그아웃"],
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/kakaologout": {
      "post": {
        "tags": ["로그인/로그아웃"],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "tags": ["로그인/로그아웃"],
        "description": "",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/googlelogout": {
      "post": {
        "tags": ["로그인/로그아웃"],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}